refactor buttons for visibility in different component, shorten code
